requirements.txt
absl-py==0.11.0
array2gif==1.0.4
astunparse==1.6.3
cached-property==1.5.2
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
classification-models-3D==1.0.2
cycler==0.10.0
decorator==4.4.2
efficientnet-3D==1.0.1
explore-lif==0.1.0
flatbuffers==1.12
gast==0.3.3
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.3
importlib-metadata==3.4.0
Keras==2.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
llvmlite==0.35.0
Markdown==3.3.3
matplotlib==3.3.4
moviepy==1.0.3
networkx==2.5
numba==0.52.0
numpy==1.18.5
oauthlib==3.1.0
opencv-python-headless==4.5.1.48
opt-einsum==3.3.0
pandas==1.1.5
pathlib2==2.3.5
Pillow==8.1.0
proglog==0.1.9
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.6.2
pyparsing==2.4.7
PyQt5==5.15.2
PyQt5-sip==12.8.1
python-dateutil==2.8.1
pytz==2020.5
PyWavelets==1.1.1
PyYAML==5.4.1
QtPy==1.9.0
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
scikit-image==0.17.2
scipy==1.4.1
segmentation-models-3D==1.0.1
six==1.15.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.3.0
tensorflow-gpu==2.3.0
termcolor==1.1.0
tifffile==2020.9.3
tqdm==4.56.0
trackpy==0.4.2
typing-extensions==3.7.4.3
urllib3==1.26.3
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0
